fixed bug where new UsernamePassword.json file was created everytime user changed database

+++ sife.py
@@ -46,9 +46,13 @@ else:
 
 def create_empty_mysql_config_file():
     destination = os.path.join(os.path.dirname(__file__), 'data/UsernamePassword.json')
+    if os.path.isfile(destination):
+        return
     file_str = '{ "username":"", "password":""}'
     with open(destination, 'w') as f:
         f.write(file_str)
+
+    return 1
 
 @click.group()
 def cli():
@@ -263,7 +267,17 @@ def set_db_engine():
         if new_db_engine_num == 1:
             new_db_engine = 'sqlite3' 
         else:
+            curr_dir = os.path.dirname(__file__)
+            config_file_path = os.path.join(curr_dir, 'data/UsernamePassword.json')
             new_db_engine = 'mysql'
+            config_created = create_empty_mysql_config_file()
+            if config_created == 1:
+                print("install the mysql-connector pip package using 'pip install mysql-connector'")
+                message_str = f"an empty config file has been generated in {config_file_path}"
+                print(colored(message_str, 'red'))
+
+
+
     else:
         new_db_engine = curr_db_engine
 
@@ -274,23 +288,13 @@ def set_db_engine():
         Migrate information from previous db to new? (m to proceed with migration).
         proceed without migrating? (Y to proceed. any other to abort)
         """)
-        curr_dir = os.path.dirname(__file__)
-        config_file_path = os.path.join(curr_dir, 'data/UsernamePassword.json')
         if allow_changing_engine == 'm':
             print(new_db_engine)
             set_dab_engine(new_db_engine)
             from utils import migrate
             migrate(curr_db_engine)
-            print("install the mysql-connector pip package using 'pip install mysql-connector'")
-            message_str = f"an empty config file has been generated in {config_file_path}"
-            print(colored(message_str, 'red'))
-            create_empty_mysql_config_file()
         elif allow_changing_engine == 'Y':
             set_dab_engine(new_db_engine)
-            print("install the mysql-connector pip package using 'pip install mysql-connector'")
-            message_str = f"an empty config file has been generated in {config_file_path}"
-            print(colored(message_str, 'red'))
-            create_empty_mysql_config_file()
         else:
             print('Abort')
